refactor(tests): move intra-package cleanup tests to their own file
Addresses review feedback (@danielmeppiel): the previous approach
moved the module-level GITHUB_APM_PAT skipif into each existing class
so the new TestFileRenamedWithinPackage could run without a token,
which meant duplicating the skipif three times.

The cleaner split: keep test_diff_aware_install_e2e.py untouched
(restored bit-for-bit from upstream/main, module-level skipif intact)
and move the new local-path-based tests to
tests/integration/test_intra_package_cleanup.py -- which has no
module-level skipif because these tests do not use GitHub at all.

No change to the tested behaviour: same fixture, same two tests.

Refs: #666
